+++ draft-packages/well/CHANGELOG.md
@@ -3,6 +3,76 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.4.2...@kaizen/draft-well@2.4.3) (2020-05-05)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.4.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.4.1...@kaizen/draft-well@2.4.2) (2020-05-05)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.4.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.4.0...@kaizen/draft-well@2.4.1) (2020-05-05)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+# [2.4.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.3.1...@kaizen/draft-well@2.4.0) (2020-05-05)
+
+
+### Features
+
+* Updated all draft packages to accept all minor releases ([#451](https://github.com/cultureamp/kaizen-design-system/issues/451)) ([80453f6](https://github.com/cultureamp/kaizen-design-system/commit/80453f6c04300dcef61c14e39200ce154863eb0d))
+
+
+
+
+
+## [2.3.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.3.0...@kaizen/draft-well@2.3.1) (2020-05-04)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+# [2.3.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.2.3...@kaizen/draft-well@2.3.0) (2020-05-04)
+
+
+### Features
+
+* Added missing dependencies to packages  ([#449](https://github.com/cultureamp/kaizen-design-system/issues/449)) ([cb1dc2a](https://github.com/cultureamp/kaizen-design-system/commit/cb1dc2aead68e591cc21c665fb25c1817633c4d7))
+
+
+
+
+
+## [2.2.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.2.2...@kaizen/draft-well@2.2.3) (2020-05-04)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.2.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.2.1...@kaizen/draft-well@2.2.2) (2020-05-04)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
 ## [2.2.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.2.0...@kaizen/draft-well@2.2.1) (2020-05-01)